# optimized frontend load time by using text compression. 