--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,4 @@
-./DSstore
+./DS_Store
 
 # Logs
 logs