--- a/package.json
+++ b/package.json
@@ -13,6 +13,7 @@
   "author": "Muizz Matemilola",
   "license": "ISC",
   "devDependencies": {
+    "@types/compression": "^1.7.2",
     "@types/express": "^4.17.13",
     "@types/mongoose-sequence": "^3.0.6",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
@@ -28,6 +29,7 @@
     "@types/cookie-parser": "^1.4.2",
     "axios": "^0.26.1",
     "bcryptjs": "^2.4.3",
+    "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",