Reductions:
- time to interactive 1.7s to 1.4s;
- total blocking 80ms to 60ms;
- largest contentful paint 1.5s to 1.2s; and
- speed index 0.7s to 0.6s.

97 performance and 86 accessibility scores on Lighthouse audit